Fixing bug
Fixing a bug in which it was giving a wrong answer when there are
repetative characters present in one string and another string has total
different charactrs.

For example one string is "aaacb" and another is "cbdef". This should
return 3 because we can either replace "aaa" of first string with "def"
of second string or "def" of second string with "aaa" of first.

Converting hash map to a Bin class, improves a readability of a class.