Improving soution
Considering Bin as a data structure to count number of characters
present in a string. A more elegant way to solve this problem is to
calculate remaining characters in bin rather than going for some
approach externally.

By silently ignoring a character which is intended to remove from a bin,
but it is not present in bin, I can reduce external code for catching
this exception at client. I am sure this can be a bad practice, but it
looks like it is working here.